refactor layout and types
