--- Testing/k6/src/application/smoke/docker-compose.yml
@@ -1,14 +1,0 @@
-version: '3.4'
-services:
-  k6:
-    image: loadimpact/k6:latest
-    command: run /application/smoke/k6_smoke.js
-    volumes:
-      - ../../../src/application/smoke:/application/smoke
-      - ../../../src/inc:/inc
-    environment: 
-      - "HG_TYPE=${HG_TYPE}"
-      - "HG_ENV=${HG_ENV}"
-      - "HG_KEY=${HG_KEY}"
-      - "HG_PASSWORD=${HG_PASSWORD}"
-
