--- Testing/k6/src/application/stress/docker-compose.yml
@@ -1,13 +1,0 @@
-version: '3.4'
-services:
-  k6:
-    image: loadimpact/k6:latest
-    command: run /application/stress/k6_stress.js
-    volumes:
-      - ../../../src/application/stress:/application/stress
-      - ../../../src/inc:/inc
-    environment: 
-      - "HG_TYPE=${HG_TYPE}"
-      - "HG_ENV=${HG_ENV}"
-      - "HG_KEY=${HG_KEY}"
-      - "HG_PASSWORD=${HG_PASSWORD}"
