+++ Testing/k6/webapp/k6_webapp.js
@@ -16,27 +16,24 @@
 
 import http from "k6/http";
 import { sleep } from "k6";
-import * as common from "../../inc/common.js";
+import * as common from "../inc/common.js";
 
-export let options = {
-    stages: [
-        { duration: "1m", target: 10 }, // below normal load
-        { duration: "2m", target: 250 },
-        { duration: "3h56m", target: 250 }, // stay at high users for hours 'soaking' the system
-        { duration: "2m", target: 0 }, // drop back down
-    ],
-};
+export let options = common.OptionConfig();
 
 export default function () {
     let user = common.users[__VU % common.users.length];
 
     common.authorizeUser(user);
 
-    let webClientBatchResponses = http.batch(common.webClientRequests(user));
-    let timelineBatchResponses = http.batch(common.timelineRequests(user));
+    common.groupWithDurationMetric("batch", function () {
+        let webClientBatchResponses = http.batch(
+            common.webClientRequests(user)
+        );
+        let timelineBatchResponses = http.batch(common.timelineRequests(user));
 
-    common.checkResponses(webClientBatchResponses);
-    common.checkResponses(timelineBatchResponses);
+        common.checkResponses(webClientBatchResponses);
+        common.checkResponses(timelineBatchResponses);
+    });
 
-    sleep(common.getRandom(1.0, 3.0));
+    sleep(1);
 }
